requirements.txt
opencv-python==4.5.5.62
numpy==1.23.5
keras==2.11.0
pyserial==3.5
keyboard==0.13.5
pyttsx3==2.90
mediapipe==0.9.0
matplotlib==3.6.0